requirements.txt
pydantic==2.11.3
pydantic-settings==2.8.1
pytest==8.3.5
pytest-picked==0.5.1
requests==2.32.3
selenium==4.31.0
faker~=24.8.0
ruff==0.11.6
allure-pytest==2.14.2
pytest-check==2.5.3
